1. Added Windows detection to Jenkinsfile. 2. Added TF files for fun. 3. Added Kubernetes deployment for fun.

+++ Jenkinsfile
@@ -28,28 +28,53 @@ pipeline {
                 }
             }
         }
-        stage('DeployToProduction') {
+        stage('Deploy To Production') {
             when {
                 branch 'master'
             }
             steps {
                 input 'Deploy to Production?'
                 milestone(1)
-                withCredentials([sshUserPrivateKey(credentialsId: 'sshUser', keyFileVariable: 'identity', usernameVariable: 'userName')]) {
-                    script {
-                        remote.user = userName
-                        remote.identityFile = identity
-                        sshCommand remote: remote, command: "docker pull shasui2/portfolio-app:${env.BUILD_NUMBER}"
-                        try {
-                            sshCommand remote: remote, command: "docker stop portfolio-app"
-                            sshCommand remote: remote, command: "docker rm portfolio-app"
-                        } catch (err) {
-                            echo: 'caught error: $err'
+                if (isUnix()) {
+                    withCredentials([sshUserPrivateKey(credentialsId: 'sshUser', keyFileVariable: 'identity', usernameVariable: 'userName')]) {
+                        script {
+                            remote.user = userName
+                            remote.identityFile = identity
+                            sshCommand remote: remote, command: "docker pull shasui2/portfolio-app:${env.BUILD_NUMBER}"
+                            try {
+                                sshCommand remote: remote, command: "docker stop portfolio-app"
+                                sshCommand remote: remote, command: "docker rm portfolio-app"
+                            } catch (err) {
+                                echo: 'caught error: $err'
+                            }
+                            sshCommand remote: remote, command: "docker run --name portfolio-app -p 80:3000 -d --network portfolio shasui2/portfolio-app:${env.BUILD_NUMBER}"
                         }
-                        sshCommand remote: remote, command: "docker run --name portfolio-app -p 3000:3000 -d --network portfolio shasui2/portfolio-app:${env.BUILD_NUMBER}"
                     }
+                }
+                else {
+                    bat 'set "You're on Windows!"'
                 }
             }
         }
     }
 }
+
+
+/*
+Test Kubernetes deployment.
+
+stage('DeployToProduction') {
+            when {
+                branch 'master'
+            }
+            steps {
+                input 'Deploy to Production?'
+                milestone(1)
+                kubernetesDeploy(
+                    kubeconfigId: 'kubeconfig',
+                    configs: 'kube-deploy.yml',
+                    enableConfigSubstitution: true
+                )
+            }
+        }
+*/
